no segfaults, errors now.

+++ tnecs.h
@@ -427,7 +427,7 @@ uint64_t tnecs_hash_sdbm(const unsigned char * str);
 #define TNECS_NAME2HASH(name) TNECS_HASH(name)
 
 // ****************** SET BIT COUNTING ************************
-int8_t setBits_KnR_uint64_t(uint64_t in_flags);
+size_t setBits_KnR_uint64_t(uint64_t in_flags);
